tg-ninja-bot-updated/telegram_bot_project/requirements.txt
python-telegram-bot==20.7
telethon==1.34.0
pyrogram==2.0.106
sqlalchemy==2.0.23
alembic==1.13.1
flask==3.0.0
flask-cors==4.0.0
apscheduler==3.10.4
openai==1.3.7
aiofiles==23.2.1
asyncio-mqtt==0.16.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0

